+++ examples/deno/wt_client.ts
@@ -11,7 +11,7 @@ const client = new WebTransport(connectAddr, {
 
 // console.log(client);
 
-await client.ready();
+await client.ready;
 console.log("Client connected");
 const writer = client.datagrams.writable.getWriter();
 writer?.write(new Uint8Array([1, 2, 3, 4, 5]));
@@ -20,7 +20,7 @@ writer?.write(new Uint8Array([1, 2, 3, 4, 5]));
 writer?.write(new Uint8Array([1, 2, 3, 4, 5]));
 writer?.write(new Uint8Array([1, 2, 3, 4, 5]));
 
-// //await messages
+// // //await messages
 for await (const read of client.datagrams.readable) {
     console.log(read);
 }
