+++ mod/client.ts
@@ -1,35 +1,33 @@
 if (import.meta.main) {
     throw new Error("This module is not meant to be imported.");
 }
-import WebTransportConnection, {
-    WebTransportDatagramDuplexStream,
-} from "./connection.ts";
-
+import WebTransportConnection from "./connection.ts";
+import { WebTransportDatagramDuplexStream } from "./streams.ts";
 import {
     type WebTransportOptions,
     WebTransportOptions as ServerOpts,
 } from "./interface.ts";
-import { encodeBuf } from "./utils.ts";
+import { decoder, encodeBuf } from "./utils.ts";
 
 export class WebTransport {
     #CONN_PTR: Deno.PointerValue<unknown> | undefined;
-    #NOTIFY_PTR = new Deno.UnsafeCallback(
-        {
-            parameters: ["u32", "pointer", "u32"],
-            result: "void",
-        },
-        this.notify.bind(this),
-    );
+    // #NOTIFY_PTR = new Deno.UnsafeCallback(
+    //     {
+    //         parameters: ["u32", "pointer", "u32"],
+    //         result: "void",
+    //     },
+    //     this.notify.bind(this),
+    // );
 
-    #CONNECTION_CB = new Deno.UnsafeCallback(
-        {
-            parameters: ["pointer"],
-            result: "void",
-        },
-        this.connection.bind(this),
-    );
+    // #CONNECTION_CB = new Deno.UnsafeCallback(
+    //     {
+    //         parameters: ["pointer"],
+    //         result: "void",
+    //     },
+    //     this.connection.bind(this),
+    // );
     public datagrams!: WebTransportDatagramDuplexStream;
-    public readonly ready: Promise<WebTransportConnection>;
+    // public readonly ready: Promise<WebTransportConnection>;
 
     public conn?: WebTransportConnection;
     protected remote: URL;
@@ -42,15 +40,10 @@ export class WebTransport {
             _client = new URL(_client);
         }
 
-        try {
-            this.#CONN_PTR = window.WTLIB.symbols.proc_client_init(
-                this.#NOTIFY_PTR.pointer,
-                _options.keepAlive,
-                _options.maxTimeout,
-            );
-        } catch (e) {
-            console.error(e);
-        }
+        this.#CONN_PTR = window.WTLIB.symbols.proc_client_init(
+            _options.keepAlive,
+            _options.maxTimeout,
+        );
 
         if (!this.#CONN_PTR) {
             throw new Error("Failed to initialize client");
@@ -59,23 +52,8 @@ export class WebTransport {
         this.remote = _client;
 
         /// ref the callback to prevent it from being garbage collected
-        this.#NOTIFY_PTR.ref();
-        this.#CONNECTION_CB.ref();
-        this.ready = new Promise<WebTransportConnection>((resolve, reject) => {
-            const encoded = encodeBuf(this.remote.href);
-            window.WTLIB.symbols.proc_client_connect(
-                this.#CONN_PTR!,
-                this.#CONNECTION_CB.pointer,
-                encoded[0],
-                encoded[1],
-            );
-            if (this.conn) {
-                resolve(this.conn);
-            } else {
-                this.conn = undefined;
-                reject("Failed to connect to server");
-            }
-        });
+        // this.#NOTIFY_PTR.ref();
+        // this.#CONNECTION_CB.ref();
     }
     /**
      * @callback connection
@@ -87,7 +65,6 @@ export class WebTransport {
         if (!client || client == null) {
             return;
         }
-        Promise.resolve(this.ready);
         const CONN_BUFFER = new Uint8Array(65536);
 
         this.conn = new WebTransportConnection(
@@ -110,26 +87,25 @@ export class WebTransport {
      *
      * @description This function is called when a new event is received from the server
      */
-    private async notify(
+    private notify(
         _code: unknown | number,
         buffer: Deno.PointerValue<unknown>,
         buflen: number,
     ) {
         const code = _code as number;
-
+        console.log("[CB CLIENT] Got code", code);
         if (buflen < 0) {
             return;
         }
-
         const pointer = Deno.UnsafePointerView.getArrayBuffer(
-            buffer as unknown as NonNullable<Deno.PointerValue>,
+            buffer!,
             buflen,
         );
-        const data = new TextDecoder().decode(pointer);
+        const data = decoder.decode(pointer);
         if (code >= 130) {
-            // await Promise.race([this.closed]);
-            console.log("We should close the connection");
-            console.log(data);
+            // Promise.race([this.closed]);
+            console.error("[CB CLIENT] We should close the connection");
+            console.error(data);
             // throw new Error(data);
         }
         const _event = new MessageEvent("error", {
@@ -138,10 +114,7 @@ export class WebTransport {
         dispatchEvent(_event);
     }
     get closed() {
-        return new Promise((resolve) => {
-            this.#NOTIFY_PTR.unref();
-            this.#CONNECTION_CB.unref();
-
+        return new Promise(() => {
             //close the datagrams
             if (this.datagrams) {
                 this.datagrams.close();
@@ -159,10 +132,27 @@ export class WebTransport {
                     this.#CONN_PTR,
                 );
             }
-
-            resolve(true);
+            // this.#NOTIFY_PTR.unref();
+            // this.#CONNECTION_CB.unref();
             // this.#NOTIFY_PTR.close();
             // this.#CONNECTION_CB.close();
+        });
+    }
+    get ready() {
+        return new Promise<WebTransportConnection>((resolve, reject) => {
+            const encoded = encodeBuf(this.remote.href);
+            const conn = window.WTLIB.symbols.proc_client_connect(
+                this.#CONN_PTR!,
+                encoded[0],
+                encoded[1],
+            );
+            this.connection(conn);
+            if (this.conn) {
+                resolve(this.conn);
+            } else {
+                this.conn = undefined;
+                reject("Failed to connect to server");
+            }
         });
     }
 }
